Fix NLP Building Bug
some courses and buildings entities share the same name but NLP always consider a an entity name as a building. To fix the bug we chaged the names of the buildings in the database that are common with the course subjects and now it works. Users can ask for buildings name to view the changes.