Fix Issue State Bug
When two users try to register an issue the bot only saves the state for one user.
This leads to only one issue submitted to the database while the other user get an error message.
Now it is fixed to remember state of several users.
Also added logging messages and unkown intents to refer to and analayze later on.